food-label-analyzer/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gunicorn>=21.2.0
sqlalchemy>=2.0.29
alembic>=1.13.1
asyncpg>=0.29.0
psycopg[binary,pool]>=3.1.18
redis>=5.0.3
celery>=5.3.6
pydantic>=2.7.0
pydantic-settings>=2.2.1
email-validator>=2.1.1
minio>=7.2.7
httpx>=0.27.0
orjson>=3.10.0
chromadb>=0.4.24
openai>=1.14.0
Pillow>=10.2.0
numpy>=1.26.0
aiosmtplib>=3.0.1
python-multipart>=0.0.18
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
structlog>=24.1.0
tenacity>=8.2.3
ultralytics>=8.1.0
onnx>=1.16.0
onnxruntime>=1.17.0
